Switch to the duct crate
Use the duct crate to fork a process, rather than Rust's
std::process::Command. This is because, we will need to capture some
output from stderr from the at(1) command. However, for whatever reason,
Rust's std::process::Command seems to be eating the line we need that
contains the job number. The duct crate seems to work correctly in this
case, when we redirect stderr to stdout.