Added facilities for getting and storing an ID
Neomutt creates unique message IDs for each email. We can use this ID
(later on) to match messages when searching through a draftbox. Added a
field in Midnight to store the given messages unique ID, and added
another function which can parse the ID when given a raw RFC 5322
message String.