Added the Draftbox model
Added a new model, Draftbox, which is able to load a .draftboxes files
from the neomutt config directory. The format of the file is as follows:

<email addr> = <path to draftbox>
...
noreply@example.com = ~/Mail/example.com/Drafts

If a .draftboxes files is found, midnight will default to using it. If a
.draftboxes file cannot be found, midnight falls back to grepping the
neomutt config dir for a relevant account file.