Disable success notifications to local mail spool
at(1), the program which midnight wraps, can be configured to send an
email to the local user account (local mail spool) upon successful
completion of the job with the '-m' flag. While this was somewhat useful
for testing early on, it doesn't seem to make much sense to do this
every time, so remove it from midnight's invocation.

Note, at(1) will still send an email to the local mail spool if the job
happens to fail, even without the '-m' flag, which is definitely good,
since it will make debugging easier, in the event a scheduled mail
doesn't send for some reason or another.